[MOVE/#3] presentation 패키지 수정
