+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation
+package org.sopt.dosopttemplate.presentation.main
 
 import android.content.Intent
 import android.content.Intent.FLAG_ACTIVITY_CLEAR_TASK
@@ -8,7 +8,8 @@ import androidx.activity.OnBackPressedCallback
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.data.model.User
 import org.sopt.dosopttemplate.databinding.ActivityMainBinding
-import org.sopt.dosopttemplate.presentation.LoginActivity.Companion.EXTRA_USER
+import org.sopt.dosopttemplate.presentation.auth.LoginActivity
+import org.sopt.dosopttemplate.presentation.auth.LoginActivity.Companion.EXTRA_USER
 import org.sopt.dosopttemplate.util.base.BindingActivity
 import org.sopt.dosopttemplate.util.intent.getParcelable
 import toast
